+++ tests/parsers/test_python.py
@@ -338,6 +338,79 @@ def test_update_python_requires_preserves_style(fix_max_python_3_version):
     """)
 
 
+def test_update_python_requires_multiline(fix_max_python_3_version):
+    fix_max_python_3_version(2)
+    fp = StringIO(textwrap.dedent("""\
+        from setuptools import setup
+        setup(
+            name='foo',
+            python_requires=', '.join([
+                '>=2.7',
+                '!=3.0.*',
+            ]),
+        )
+    """))
+    fp.name = "setup.py"
+    result = update_python_requires(fp, ['2.7', '3.2'])
+    assert "".join(result) == textwrap.dedent("""\
+        from setuptools import setup
+        setup(
+            name='foo',
+            python_requires=', '.join([
+                '>=2.7',
+                '!=3.0.*',
+                '!=3.1.*',
+            ]),
+        )
+    """)
+
+
+def test_update_python_requires_multiline_variations(fix_max_python_3_version):
+    fix_max_python_3_version(2)
+    fp = StringIO(textwrap.dedent("""\
+        from setuptools import setup
+        setup(
+            name='foo',
+            python_requires=",".join([
+                ">=2.7",
+                "!=3.0.*",
+            ]),
+        )
+    """))
+    fp.name = "setup.py"
+    result = update_python_requires(fp, ['2.7', '3.2'])
+    assert "".join(result) == textwrap.dedent("""\
+        from setuptools import setup
+        setup(
+            name='foo',
+            python_requires=",".join([
+                ">=2.7",
+                "!=3.0.*",
+                "!=3.1.*",
+            ]),
+        )
+    """)
+
+
+def test_update_python_requires_multiline_error(capsys):
+    fp = StringIO(textwrap.dedent("""\
+        from setuptools import setup
+        setup(
+            name='foo',
+            python_requires=', '.join([
+                '>=2.7',
+                '!=3.0.*']),
+        )
+    """))
+    fp.name = "setup.py"
+    result = update_python_requires(fp, ['2.7', '3.2'])
+    assert result == fp.getvalue().splitlines(True)
+    assert (
+        "Did not understand python_requires= formatting in setup() call"
+        in capsys.readouterr().err
+    )
+
+
 def test_find_call_kwarg_in_ast():
     tree = ast.parse('foo(bar="foo")')
     ast.dump(tree)